add embeddings_openai.py utility
The purpose of this utility is to generate asynchronously vector embeddings of the output of the ocr_openai.py utility, which thereafter will imported to a vector database. The latter is a feature that will be integrated in the upcoming pushes. It includes many of the auxiliary features of the utility ocr_openai.py, as well as others specifically for the OpenAI Vector Embeddings API, such as configurable dimensionality for the supported models and configurable encoding format.